Solve #33 and .gitattributes

+++ src/utils/trash.sh
@@ -5,8 +5,14 @@ source $(dirname "$(readlink -f "$0")")/utils/global.sh
 if [[ -d ~/.local/share/Trash/files ]]; then
 	printf "$RED\nShowing files in .local/share/Trash/files$NORMAL\n"
 	ls -Ahl ~/.local/share/Trash/files
-	sudo rm -rf ~/.local/share/Trash/*
-	printProgress "Trash: cleaned"
+	printf "Should I clean Trash? "
+	read -p "[Y/n]: " ANSW
+	if [[ "$ANSW" == "y" ]]; then
+	    sudo rm -rf ~/.local/share/Trash/*
+	    printProgress "Trash: cleaned"
+	else
+	    printProgress "Trash: not cleaned"
+	fi
 else
     printProgress "\nTrash is empty, nothing to clean"
 fi
